Refactor xtask tooling into modular commands
