+++ xtask/src/commands/docs.rs
@@ -1,0 +1,114 @@
+use crate::error::{TaskError, TaskResult};
+use crate::util::{is_help_flag, run_cargo_tool};
+use std::ffi::OsString;
+use std::path::Path;
+
+/// Options accepted by the `docs` command.
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub struct DocsOptions {
+    /// Whether to open the generated documentation in a browser.
+    pub open: bool,
+}
+
+/// Parses CLI arguments for the `docs` command.
+pub fn parse_args<I>(args: I) -> TaskResult<DocsOptions>
+where
+    I: IntoIterator<Item = OsString>,
+{
+    let mut options = DocsOptions::default();
+
+    for arg in args {
+        if is_help_flag(&arg) {
+            return Err(TaskError::Help(usage()));
+        }
+
+        if arg == "--open" {
+            if options.open {
+                return Err(TaskError::Usage(String::from(
+                    "--open specified multiple times",
+                )));
+            }
+            options.open = true;
+            continue;
+        }
+
+        return Err(TaskError::Usage(format!(
+            "unrecognised argument '{}' for docs command",
+            arg.to_string_lossy()
+        )));
+    }
+
+    Ok(options)
+}
+
+/// Executes the `docs` command.
+pub fn execute(workspace: &Path, options: DocsOptions) -> TaskResult<()> {
+    println!("Building API documentation");
+    let mut doc_args = vec![
+        OsString::from("doc"),
+        OsString::from("--workspace"),
+        OsString::from("--no-deps"),
+        OsString::from("--locked"),
+    ];
+    if options.open {
+        doc_args.push(OsString::from("--open"));
+    }
+
+    run_cargo_tool(
+        workspace,
+        doc_args,
+        "cargo doc",
+        "ensure the Rust toolchain is installed",
+    )?;
+
+    println!("Running doctests");
+    let test_args = vec![
+        OsString::from("test"),
+        OsString::from("--doc"),
+        OsString::from("--workspace"),
+        OsString::from("--locked"),
+    ];
+
+    run_cargo_tool(
+        workspace,
+        test_args,
+        "cargo test --doc",
+        "ensure the Rust toolchain is installed",
+    )
+}
+
+/// Returns usage text for the command.
+pub fn usage() -> String {
+    String::from(
+        "Usage: cargo xtask docs [--open]\n\nOptions:\n  --open          Open documentation after building\n  -h, --help      Show this help message",
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn parse_args_accepts_default_configuration() {
+        let options = parse_args(std::iter::empty()).expect("parse succeeds");
+        assert_eq!(options, DocsOptions::default());
+    }
+
+    #[test]
+    fn parse_args_reports_help_request() {
+        let error = parse_args([OsString::from("--help")]).unwrap_err();
+        assert!(matches!(error, TaskError::Help(message) if message == usage()));
+    }
+
+    #[test]
+    fn parse_args_rejects_duplicate_open_flags() {
+        let error = parse_args([OsString::from("--open"), OsString::from("--open")]).unwrap_err();
+        assert!(matches!(error, TaskError::Usage(message) if message.contains("--open")));
+    }
+
+    #[test]
+    fn parse_args_rejects_unknown_argument() {
+        let error = parse_args([OsString::from("--unknown")]).unwrap_err();
+        assert!(matches!(error, TaskError::Usage(message) if message.contains("--unknown")));
+    }
+}
